Rename charts.ts to packages.ts
