+++ dashboard/src/actions/index.ts
@@ -2,7 +2,7 @@ import { push } from "connected-react-router";
 import * as apps from "./apps";
 import * as auth from "./auth";
 import * as catalog from "./catalog";
-import * as charts from "./charts";
+import * as charts from "./packages";
 import * as config from "./config";
 import * as kube from "./kube";
 import * as namespace from "./namespace";
